Changed the main module to better support running in Docker
